requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
firebase-admin==6.2.0
openai==0.28.1
python-dotenv==1.0.0
